Skip some checks if we already point to the destination release
The following scenario created an issue:

- perform all steps from phase 1 and 2
- exit the application
- rerun the application

The first time the upgrader is run it makes the OS point to the new
release's repositories, but because we exited the app, the cache
isn't upgraded.

Running apt orphan checks again creates mayhem and wants to remove
packages.

This commit checks if we're pointing to the new release. If we are,
we skip these checks in phase 1.