Fix orphan management
Separate the two checks (pre and post-upgrade).

In the first one we simply show the user orphan packages. We can
keep some if he wants.

We save the list for later.

In the second one, post-upgrade, we detect orphans again and detect
the ones which were not in the initial list. These newly orphaned pkgs
can then be removed without user interaction.

We only run these steps if the upgrade starts with APT pointing at
the origin, otherwise we've no way of knowing if an orphan is 
something the user added, or something that the new base no longer
needs.