Add celluloid, simple-scan, hypnotix
Because of the t64 transition between 22.04 and 24.04
some of the lib packages become newly foreign packages but they're
provided by t64 pkgs in 24.04... and these apps depend on the virtual
name.

So when these pkgs are removed, it removes these apps with them.

Re-add these apps at the end of the upgrade.